Module atmega32u4_hal::pac

Reexport of atmega32u4 from avr-device ATmega32U4 Peripheral access API for ATMEGA32U4 microcontrollers (generated using svd2rust v0.17.0 (29129c0e4e26 2020-09-23))

You can find an overview of the generated API here.

API features to be included in the next svd2rust release can be generated by cloning the svd2rust repository, checking out the above commit, and running cargo doc --open.

Modules

ac
Analog Comparator

adc
Analog-to-Digital Converter

boot_load
Bootloader

cpu
CPU Registers

eeprom
EEPROM

exint
External Interrupts

fuse
Fuses

jtag
JTAG Interface

lockbit
Lockbits

pll
Phase Locked Loop

portb
I/O Port

portc
I/O Port

portd
I/O Port

porte
I/O Port

portf
I/O Port

spi
Serial Peripheral Interface

tc0
Timer/Counter, 8-bit

tc1
Timer/Counter, 16-bit

tc3
Timer/Counter, 16-bit

tc4
Timer/Counter, 10-bit

twi
Two Wire Serial Interface

usart1
USART

usb_device
USB Device Registers

wdt
Watchdog Timer

Structs

AC
Analog Comparator

ADC
Analog-to-Digital Converter

BOOT_LOAD
Bootloader

CPU
CPU Registers

EEPROM
EEPROM

EXINT
External Interrupts

FUSE
Fuses

JTAG
JTAG Interface

LOCKBIT
Lockbits

PLL
Phase Locked Loop

PORTB
I/O Port

PORTC
I/O Port

PORTD
I/O Port

PORTE
I/O Port

PORTF
I/O Port

Peripherals
All the peripherals

SPI
Serial Peripheral Interface

TC0
Timer/Counter, 8-bit

TC1
Timer/Counter, 16-bit

TC3
Timer/Counter, 16-bit

TC4
Timer/Counter, 10-bit

TWI
Two Wire Serial Interface

USART1
USART

USB_DEVICE
USB Device Registers

WDT
Watchdog Timer

Enums

Interrupt
Enumeration of all the interrupts